Revert "Add ping column (#9)" (#13)
The authors of the plugin hub to dot want two plugins pinging the osrs
servers (understandably). The correct way to do this would be to expose
the pings that the world hopper plugin collects through an api, and use
those pings. Until that happens, we will revert this feature, and can
revisit at a later date.

This reverts commit 38c13f299aa8065943baafad502cbc3093222402.

Co-authored-by: JC Arbelbide <jc.arbelbide@truefootage.tech>